refactor: update all refs to src --> components
