+++ .github/workflows/cdn_deploy.yml
@@ -8,7 +8,7 @@ on:
       - "*"
     # Workflow is triggered only if src changes
     paths:
-      - src/**
+      - components/**
   # Allow manual trigger (workflow_dispatch)
   workflow_dispatch:
