+++ .github/workflows/docs.yml
@@ -4,7 +4,7 @@ on:
   push:
     paths:
       - docs/**
-      - src/**
+      - components/**
       - mkdocs.yml
     branches: [main]
   # Allow manual trigger (workflow_dispatch)
