+++ tsconfig.json
@@ -10,7 +10,7 @@
     "outDir": "dist",
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"]
+      "@/*": ["components/*"]
     },
     "pretty": true,
     "newLine": "lf",
@@ -25,7 +25,7 @@
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": ["src"],
+  "include": ["components"],
   "ts-node": {
     "transpileOnly": true,
     "files": true,
