+++ tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   clean: true,
   sourcemap: true,
-  entry: ["src/index.ts"],
+  entry: ["components/index.ts"],
   format: ["cjs", "esm"],
   dts: true,
   target: "node16",
