+++ vite.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
   // base: "",
   build: {
     lib: {
-      entry: resolve(__dirname, "src/index.ts"),
+      entry: resolve(__dirname, "components/index.ts"),
       name: "hot",
     },
   },
